Keep process-detail helpers private
Greptile's last review found that extracting process-detail helpers widened private snapshot construction utilities to module-internal access. The helper calls are only needed by the private processInfo path, so keep them private in the snapshot file and move the already-internal cmux scope parsing beside the scope cache that consumes it.

Constraint: Do not touch YAML files while updating this PR

Rejected: Leave helpers internal | preserves an avoidable API-surface expansion

Confidence: high

Scope-risk: narrow

Directive: Keep low-level proc/task decoding helpers scoped to snapshot construction unless a real external caller appears

Tested: swiftc -default-isolation MainActor -typecheck Sources/CmuxTopSnapshot.swift Sources/CmuxTopSnapshotScopeCache.swift Sources/CmuxTopProcessCPUTracker.swift Sources/CmuxTopProcessDetails.swift

Tested: swiftc -typecheck Sources/CmuxTopSnapshot.swift Sources/CmuxTopSnapshotScopeCache.swift Sources/CmuxTopProcessCPUTracker.swift Sources/CmuxTopProcessDetails.swift

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: git diff --check
